change ktap_vaule n field to "long" instead of "int"
In x86-64 Linux, int is just 4 bytes, which cannot respesent
ip vaule(8 bytes).

we also can use "void *" to respesent ip, but then we need
to use a new type like TYPE_IP, currently we don't use that
method.

This patch also add some kmalloc failure return check.

+++ include/ktap_types.h
@@ -87,7 +87,7 @@ union _ktap_value {
 	void *p;         /* light userdata */
 	int b;           /* booleans */
 	ktap_cfunction f; /* light C functions */
-	int n;         /* numbers */
+	long n;         /* numbers */
 };
